--- service-post/src/main/java/com/team13/servicepost/controller/LikePostController.java
@@ -1,31 +1,0 @@
-package com.team13.servicepost.controller;
-
-import com.team13.servicepost.service.LikePostService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-
-@RestController
-@RequestMapping("/posts")
-@RequiredArgsConstructor
-public class LikePostController {
-
-    private final LikePostService likePostService;
-
-    @PostMapping("/{postId}/like")
-    public ResponseEntity<String> toggleLikePost(@PathVariable Long postId, @RequestParam Long userId) {
-        boolean success = likePostService.toggleLikePost(postId, userId);
-        if (success) {
-            return ResponseEntity.ok("Post liked/unliked successfully.");
-        } else {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body("Operation failed: Post not found or user not found.");
-        }
-    }
-
-    @GetMapping("/{postId}/likes")
-    public ResponseEntity<Long> getLikesCount(@PathVariable Long postId) {
-        Long likesCount = likePostService.getLikesCount(postId);
-        return ResponseEntity.ok(likesCount);
-    }
-}
